Move social auth state to central auth state
