--- src/containers/SocialAuth/constants.js
@@ -1,5 +1,0 @@
-export const SOCIAL_AUTH = '[Social Authentication]';
-
-export const SOCIAL_AUTH_REQUEST = `${SOCIAL_AUTH} SOCIAL_AUTH_REQUEST`;
-export const SOCIAL_AUTH_SUCCESS = `${SOCIAL_AUTH} SOCIAL_AUTH_SUCCESS`;
-export const SOCIAL_AUTH_ERROR = `${SOCIAL_AUTH} SOCIAL_AUTH_ERROR`;
