+++ src/store/auth/actionTypes.js
@@ -18,3 +18,6 @@ export const REGISTER_ERROR = `${AUTH} REGISTER_ERROR`;
 export const RESET_PASSWORD_REQUEST = `${AUTH} RESET_PASSWORD_REQUEST`;
 export const RESET_PASSWORD_SUCCESS = `${AUTH} RESET_PASSWORD_SUCCESS`;
 export const RESET_PASSWORD_ERROR = `${AUTH} RESET_PASSWORD_ERROR`;
+export const SOCIAL_AUTH_REQUEST = `${AUTH} SOCIAL_AUTH_REQUEST`;
+export const SOCIAL_AUTH_SUCCESS = `${AUTH} SOCIAL_AUTH_SUCCESS`;
+export const SOCIAL_AUTH_ERROR = `${AUTH} SOCIAL_AUTH_ERROR`;
